Corrige integracion del modulo de favoritos

+++ back-ecommerce/src/main/java/com/uade/ecommerce/dto/FavoritoDTO.java
@@ -1,10 +1,12 @@
 package com.uade.ecommerce.dto;
 
-import com.uade.ecommerce.entity.Favorito; // O com.uade.ecommerce.model.Favorito según el paquete de tus entidades
+import com.uade.ecommerce.model.Favorito;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -15,19 +17,18 @@ public class FavoritoDTO {
     private Long id;
     private Long usuarioId;
     private Long productoId;
-    
     private String nombreProducto;
-    private Double precioProducto;
-    private Boolean stockDisponible; 
+    private BigDecimal precioProducto;
+    private Boolean stockDisponible;
 
     public static FavoritoDTO fromEntity(Favorito favorito) {
         return new FavoritoDTO(
                 favorito.getId(),
                 favorito.getUsuario().getId(),
                 favorito.getProducto().getId(),
-                favorito.getProducto().getNombre(), 
+                favorito.getProducto().getNombre(),
                 favorito.getProducto().getPrecio(),
-                favorito.getProducto().getStock() > 0 
+                favorito.getProducto().getStock() > 0
         );
     }
 }
